Updated hooks
