+++ web/modules/custom/hoeringsportal_activity/hoeringsportal_activity.services.yml
@@ -3,3 +3,5 @@ services:
     autowire: true
 
   Drupal\hoeringsportal_activity\Helper\ActivityHelper:
+
+  Drupal\hoeringsportal_activity\Hook\ActivityHooks:
