+++ web/modules/custom/hoeringsportal_project/hoeringsportal_project.services.yml
@@ -2,4 +2,4 @@ services:
   _defaults:
     autowire: true
 
-  Drupal\hoeringsportal_project\Helper\ProjectHelper:
+  Drupal\hoeringsportal_project\Hook\ProjectHooks:
