+++ web/modules/custom/hoeringsportal_project/src/Hook/ProjectHooks.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\hoeringsportal_project\Helper;
+namespace Drupal\hoeringsportal_project\Hook;
 
 use Drupal\Core\Datetime\DrupalDateTime;
 use Drupal\Core\Entity\EntityInterface;
@@ -18,9 +18,9 @@ use Drupal\Core\Logger\LoggerChannelFactoryInterface;
 use Drupal\Core\Hook\Attribute\Hook;
 
 /**
- * Helper class for project-related operations.
+ * Hooks for project-related operations.
  */
-class ProjectHelper {
+class ProjectHooks {
   use StringTranslationTrait;
 
   /**
@@ -44,8 +44,8 @@ class ProjectHelper {
    * @param array $variables
    *   The template variables array.
    */
-  #[Hook('preprocess')]
-  public function projectPreprocess(array &$variables): void {
+  #[Hook('preprocess_node')]
+  public function preprocessNode(array &$variables): void {
     if ('full' === $variables['view_mode'] && 'project_main_page' === $variables['node']->bundle()) {
       if (!$variables['node']->field_show_timeline->value) {
         return;
