+++ web/sites/development.services.yml
@@ -19,12 +19,12 @@ services:
     class: Drupal\Core\Cache\NullBackendFactory
   logger.channel.config_schema:
     parent: logger.channel_base
-    arguments: ["config_schema"]
+    arguments: [ 'config_schema' ]
   config.schema_checker:
     class: Drupal\Core\Config\Development\LenientConfigSchemaChecker
     arguments:
-      - "@config.typed"
-      - "@messenger"
-      - "@logger.channel.config_schema"
+      - '@config.typed'
+      - '@messenger'
+      - '@logger.channel.config_schema'
     tags:
       - { name: event_subscriber }
